// frontend/src/lib/api/models/PrincipalLeaderResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PrincipalLeaderResponseDto = {
    id: string;
    role: string;
    name: string;
    photoUrl?: string | null;
    displayOrder: number;
    isActive: boolean;
};