// frontend/src/lib/api/models/MinistryLeadershipDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MinistryLeadershipDto = {
    id: string;
    name: string;
    sigla?: string | null;
    color: string;
    /**
     * Nombres de los responsables/directores, unidos por " y ".
     */
    leaders?: string | null;
};